func_new_car/requirements.txt
flask==2.3.3
google-cloud-firestore
numpy
google-cloud-pubsub
